feat(Dealer): Player의 승패 여부 판단 로직을 Dealer로 이동
- Referee, RefereeTest 삭제
- Referee 내 Player의 승패 여부 판단 로직을 Dealer로 이동
- RefereeTest 내 Player의 승패 여부 판단 테스트 로직을 DealerTest로 이동
