+++ README.md
@@ -63,6 +63,7 @@
 - [X] 딜러의 이름은 항상 "딜러" 이다.
 - [X] 딜러는 자신의 카드들을 가지고 있다.
 - [X] 카드의 상태를 알아낸다. (21: Black Jack, 21 초과: BURST, 16 초과: STAY, 16 이하: HIT)
+- [X] 딜러 자신의 카드 패의 합과 플레이어들의 카드 패의 합을 각각 비교해서 플레이어들의 승패 여부를 판단한다.
 
 ### 블랙잭 게임
 
@@ -81,10 +82,6 @@
 ### WinningResultStatus
 - [X] 승리, 패배, 무승부를 저장하는 enum
 - [X] 플레이어의 승리는 딜러의 패배로 플레이어의 패배는 딜러의 승리로 기록된다.
-
-### Referee
-
-- [X] 딜러 카드 패의 합과 플레이어들의 카드 패의 합을 각각 비교해서 플레이어들의 승패 여부를 판단한다.
 
 ###  InputView
 - [X] 게임에 참여할 사람의 이름을 입력한다.
